requirements.txt
streamlit
Pillow
requests
python-dotenv
gcsfs
# Direct pypi install coming soon
git+https://github.com/streamlit/files-connection
google-cloud-vision
streamlit-draggable-list
